refactor(Lottos): 정적 팩토리 메서드명이 역할을 잘 드러내도록 변경한다
